feat: stores for all data fetching, real-time orders, decimal ratings
Stores / data layer:
- Move all direct Prisma calls in account pages to client stores
- reviewStore: add fetchReviews(productId) + refetch after submit
- contentStore: add fetchContent() replacing inline fetch in ContentSection
- addressStore: add addresses state + fetchAddresses() replacing inline fetch
- accountStore (new): fetch /api/account for profile + stats
- orderStore: add myOrders + fetchMyOrders() + applyOrderUpdate() for SSE

API routes:
- GET /api/orders — customer's own orders
- GET /api/account — user profile + order/wishlist counts
- GET /api/orders/stream — SSE endpoint, polls every 8s and streams status updates to authenticated customer

Pages converted (server Prisma → client store):
- account/page.tsx
- account/orders/page.tsx (with SSE EventSource subscription)
- account/addresses/page.tsx

Components updated:
- ProductReviews: use reviewStore, render half-stars
- ContentSection: use contentStore.fetchContent()
- ShippingForm: use addressStore instead of inline fetch

Decimal ratings:
- prisma schema: rating Int → Float
- migration SQL for DOUBLE PRECISION cast
- reviews API: parseInt → parseFloat
- ReviewModal: half-star picker (0.5 increments, hover + click)
- ProductReviews: StarDisplay with half-fill overlay

+++ src/app/api/reviews/post.ts
@@ -21,7 +21,7 @@ export default async function handler(req: NextRequest & { userEmail?: string })
         userId: user.id,
         productId,
         orderId: orderId || null,
-        rating: parseInt(rating),
+        rating: parseFloat(rating),
         comment,
       }
     });
